+++ src/pytorch_lightning/utilities/__init__.py
@@ -23,7 +23,6 @@ from pytorch_lightning.utilities.enums import GradClipAlgorithmType  # noqa: F40
 from pytorch_lightning.utilities.grads import grad_norm  # noqa: F401
 from pytorch_lightning.utilities.imports import (  # noqa: F401
     _HIVEMIND_AVAILABLE,
-    _HOROVOD_AVAILABLE,
     _HPU_AVAILABLE,
     _IPU_AVAILABLE,
     _OMEGACONF_AVAILABLE,